cert-renewal-agent/requirements.txt
PyYAML>=6.0
cryptography>=41.0
requests>=2.31
